src/requirements.txt
Django==1.8
Pillow==3.0.0
django-mysql==0.2.3
djangorestframework==3.3.0
mysqlclient==1.3.6
